+++ go.mod
@@ -1,10 +1,13 @@
-module GoDuck
+module github.com/codenym/goduck
 
 go 1.21.5
 
 require (
 	github.com/aws/aws-sdk-go v1.50.5 // indirect
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/jmespath/go-jmespath v0.4.0 // indirect
 	github.com/marcboeker/go-duckdb v1.5.6 // indirect
 	github.com/mitchellh/mapstructure v1.5.0 // indirect
+	github.com/spf13/cobra v1.8.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
 )
